changes in model and val function
