+++ openfl-tutorials/interactive_api/PyTorch_Lightning_MNIST_GAN/envoy/envoy_config_two.yaml
@@ -1,0 +1,9 @@
+params:
+  cuda_devices: [1,3]
+
+optional_plugin_components: {}
+
+shard_descriptor:
+  template: mnist_shard_descriptor.MnistShardDescriptor
+  params:
+    rank_worldsize: 2, 2
