+++ openfl-tutorials/interactive_api/PyTorch_Lightning_MNIST_GAN/envoy/sd_requirements.txt
@@ -1,2 +1,7 @@
 numpy
 pillow
+pytorch-lightning>=1.3
+torch==1.9.1
+torchvision==0.10.1
+torchmetrics>=0.3
+tensorflow=2.7.0
