+++ openfl-tutorials/interactive_api/PyTorch_Lightning_MNIST_GAN/envoy/start_envoy_two.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+fx envoy start -n env_two --disable-tls --envoy-config-path envoy_config_two.yaml -dh localhost -dp 50050
